Move some misc stuff from game/common to map_gui, so LTN can use it
LTN crate builds, and appears to work!
