+++ game/src/edit/zones.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeSet;
 use enumset::EnumSet;
 use maplit::btreeset;
 
-use map_gui::tools::ColorDiscrete;
+use map_gui::tools::{intersections_from_roads, ColorDiscrete};
 use map_model::{AccessRestrictions, CommonEndpoint, PathConstraints, RoadID};
 use sim::TripMode;
 use widgetry::mapspace::ToggleZoomed;
@@ -14,7 +14,7 @@ use widgetry::{
 
 use crate::app::{App, Transition};
 use crate::common::RoadSelector;
-use crate::common::{checkbox_per_mode, intersections_from_roads, CommonState};
+use crate::common::{checkbox_per_mode, CommonState};
 use crate::edit::apply_map_edits;
 
 pub struct ZoneEditor {
