+++ map_gui/src/tools/ui.rs
@@ -1,11 +1,13 @@
 //! Generic UI tools. Some of this should perhaps be lifted to widgetry.
 
+use std::cmp::Ordering;
+
 use anyhow::Result;
 
-use geom::Polygon;
+use geom::{Distance, Duration, Polygon};
 use widgetry::{
-    hotkeys, Choice, Color, DrawBaselayer, EventCtx, GfxCtx, Key, Line, Menu, Outcome, Panel,
-    State, Text, TextBox, Transition, Widget,
+    hotkeys, Choice, Color, DrawBaselayer, EventCtx, GeomBatch, GfxCtx, Key, Line, Menu, Outcome,
+    Panel, State, Text, TextBox, Transition, Widget,
 };
 
 use crate::load::FutureLoader;
@@ -232,5 +234,85 @@ impl FilePicker {
             "Waiting for a file to be chosen",
             on_load,
         )
+    }
+}
+
+pub fn percentage_bar(ctx: &EventCtx, txt: Text, pct_green: f64) -> Widget {
+    let bad_color = Color::RED;
+    let good_color = Color::GREEN;
+
+    let total_width = 450.0;
+    let height = 32.0;
+    let radius = 4.0;
+
+    let mut batch = GeomBatch::new();
+    // Background
+    batch.push(
+        bad_color,
+        Polygon::rounded_rectangle(total_width, height, radius),
+    );
+    // Foreground
+    if let Some(poly) = Polygon::maybe_rounded_rectangle(pct_green * total_width, height, radius) {
+        batch.push(good_color, poly);
+    }
+    // Text
+    let label = txt.render_autocropped(ctx);
+    let dims = label.get_dims();
+    batch.append(label.translate(10.0, height / 2.0 - dims.height / 2.0));
+    batch.into_widget(ctx)
+}
+
+/// Shorter is better
+pub fn cmp_dist(txt: &mut Text, app: &dyn AppLike, dist: Distance, shorter: &str, longer: &str) {
+    match dist.cmp(&Distance::ZERO) {
+        Ordering::Less => {
+            txt.add_line(
+                Line(format!(
+                    "{} {}",
+                    (-dist).to_string(&app.opts().units),
+                    shorter
+                ))
+                .fg(Color::GREEN),
+            );
+        }
+        Ordering::Greater => {
+            txt.add_line(
+                Line(format!("{} {}", dist.to_string(&app.opts().units), longer)).fg(Color::RED),
+            );
+        }
+        Ordering::Equal => {}
+    }
+}
+
+/// Shorter is better
+pub fn cmp_duration(
+    txt: &mut Text,
+    app: &dyn AppLike,
+    duration: Duration,
+    shorter: &str,
+    longer: &str,
+) {
+    match duration.cmp(&Duration::ZERO) {
+        Ordering::Less => {
+            txt.add_line(
+                Line(format!(
+                    "{} {}",
+                    (-duration).to_string(&app.opts().units),
+                    shorter
+                ))
+                .fg(Color::GREEN),
+            );
+        }
+        Ordering::Greater => {
+            txt.add_line(
+                Line(format!(
+                    "{} {}",
+                    duration.to_string(&app.opts().units),
+                    longer
+                ))
+                .fg(Color::RED),
+            );
+        }
+        Ordering::Equal => {}
     }
 }
